+++ src/moduls/daq/JavaLikeCalc/virtual.cpp
@@ -497,7 +497,7 @@ void Contr::start_( )
     if(id_this >= 0) setO(id_this,new TCntrNodeObj(AutoHD<TCntrNode>(this),"root"));
 
     //> Schedule process
-    mPer = TSYS::strSepParse(mSched,1,' ').empty() ? vmax(0,(long long)(1e9*atof(mSched.getVal().c_str()))) : 0;
+    mPer = TSYS::strSepParse(mSched,1,' ').empty() ? vmax(0,(int64_t)(1e9*atof(mSched.getVal().c_str()))) : 0;
 
     //> Start the request data task
     if(!prc_st) SYS->taskCreate(nodePath('.',true), mPrior, Contr::Task, this, &prc_st);
